% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alstyne}
\alias{alstyne}
\title{van Alstyne and Gottfredson Data}
\format{
This is a \code{data.frame} object, with the following
variables:
\describe{
  \item{person}{Was this a crime against a person?}
  \item{age}{Was the perpatrator aged over 25?}
  \item{drug}{Did they have a drug of alcohol dependancy?}
  \item{validation}{Indictator that this was from a validation sample}
  \item{success}{Was the parolee successful in avoiding arrest for further offences?}
  \item{prior}{Was there a previous sentence for this offence?}
  \item{freq}{The total number of individuals in this row}
}
}
\source{
van Alstyne and Gottfredson, A Multidimensional Contingency Table Analysis of Parole Outcome: New Methods and Old Problems in Criminological Prediction,
\emph{J. Res. Crime and Deliquency}, 1978.
}
\usage{
alstyne
}
\description{
A criminology dataset for illustration.
}
\keyword{datasets}
