% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{ancDisClos}
\alias{ancDisClos}
\title{Get C^*}
\usage{
ancDisClos(graph, C, int)
}
\arguments{
\item{graph}{ADMG object of class \code{mixedgraph}}

\item{C}{\code{intrinsic set}}

\item{int}{optional list of intrinsic sets}
}
\description{
Get C^*
}
\details{
Obtains largest intrinsic set of which the 
intrinsic set \code{C} is an ancestral district.  If \code{C} is
an ancestral district, then the code just returns 
the vertices of \code{graph}.
}
