% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arid.R
\name{aridProj}
\alias{aridProj}
\title{Compute the (maximal) arid projection of a graph}
\usage{
aridProj(graph, maximal = TRUE, verbose = FALSE)
}
\arguments{
\item{graph}{an ADMG as an object of class \code{mixedgraph}}

\item{maximal}{should the projection computed be maximal?}

\item{verbose}{logical: should additional output be given?}
}
\description{
Compute the (maximal) arid projection of a graph
}
\details{
Algorithm: compute the intrinsic closure of every 
vertex. Use this to obtain directed edges. Then add in bidirected
edges for vertices not already adjacent.  Then go through pairs
and check if intrinsic closures are joined by a bidirected edge.
If maximal graph asked for, then also check intrinsic closures
of the pairs.
}
