% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{as.ci}
\alias{as.ci}
\title{Turn list into a conditional independence}
\usage{
as.ci(x, ...)
}
\arguments{
\item{x}{either a list with two or three elements, or a vector}

\item{...}{if \code{x} not a list, one or two more arguments}
}
\description{
Turn list into a conditional independence
}
\examples{
as.ci(list(1,2))
as.ci(list(1,2,3))
as.ci(1,2)
as.ci(1,2,3)

}
