% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFit.R
\name{autoFit}
\alias{autoFit}
\title{Find best ADMG model to data.}
\usage{
autoFit(
  dat,
  start = NULL,
  criterion = "AIC",
  quietly = FALSE,
  r = TRUE,
  search = "basic",
  tol = 0.01
)
}
\arguments{
\item{dat}{Either an object of class \code{freq.table}, or a vector of
counts in reverse lexicographical order.}

\item{start}{Graph to start search procedure at; defaults to graph with no
edges.}

\item{criterion}{Criterion to minimize: either \code{"AIC"} or \code{"BIC"}.}

\item{quietly}{Logical indicating whether to suppress output.}

\item{r}{Logical indicating whether recursive factorizations should be used.}

\item{search}{\code{"full"} or \code{"basic"} search.}

\item{tol}{tolerance used to assess convergence in \code{\link{fitADMG}}.}
}
\value{
An object of class \code{mixed_fit}.
}
\description{
Fit discrete data to the Markov structure implied by an acyclic directed
mixed graph.
}
\details{
Given the current graph, the algorithm fits all graphs obtained by removing,
adding or (in the case of the full search) replacing a single edge, and
picks the graph which is optimal with respect to the chosen criterion (AIC
or BIC).  The algorithm terminates when no move results in an improvement.

The default starting graph has no edges.
}
\note{
Note that this procedure is not guaranteed to find the best model fit
for all ADMGs, but will find a 'local maximum' with respect to edge addition
and removal.
}
\examples{

data(gss.small)
\dontrun{out = autoFit(gss.small)}

\dontrun{summary(out$graph)}

}
\references{
Evans and Richardson (2009) - Maximum likelihood fitting of
acyclic directed mixed graphs to binary data.
}
\seealso{
\code{\link{fitADMG}}.
}
\author{
Ilya Shpitser, Robin Evans
}
\keyword{graphs}
