% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caesarian}
\alias{caesarian}
\title{Caesarian Dataset}
\format{This is a \code{data.frame} object, with the following
variables:
\describe{
  \item{cesarian}{Was a Caesarian section performed?}
  \item{monitor}{}
  \item{arrest}{}
  \item{breech}{Was the baby in the breech position?}
  \item{nullipar}{Is this a first birth?}
  \item{freq}{The total number of births in this row}
}}
\source{
Unknown.
}
\usage{
caesarian
}
\description{
A dataset on Caesarian sections for illustration.
}
\keyword{datasets}
