% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{ci_operations}
\alias{ci_operations}
\alias{split_ci}
\alias{standardize_cis}
\title{Operations for conditional independences}
\usage{
split_ci(cis)

standardize_cis(cis)
}
\arguments{
\item{cis}{object of class \code{ci} or list thereof}
}
\value{
A list with the relevant output in, one entry for each conditional 
independence put into \code{cis}.
}
\description{
Operations on conditional independence objects
}
\section{Functions}{
\itemize{
\item \code{split_ci}: Split conditional independences into univariate pieces

\item \code{standardize_cis}: Standardize conditional independences
}}

