% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{constr_sets}
\alias{constr_sets}
\title{Get sets constrained by a conditional independence}
\usage{
constr_sets(ci, sort = 1, max = Inf, check = TRUE)
}
\arguments{
\item{ci}{a conditional independence or list of \code{ci} objects}

\item{check}{logical: should validity of \code{ci} be checked?}
}
\description{
Get sets constrained by a conditional independence
}
