% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sep.R
\name{findSep}
\alias{findSep}
\alias{listSep}
\alias{allSep}
\alias{whSep}
\title{Find separator sets}
\usage{
findSep(graph, X, Y, Z = integer(0), R, check = TRUE)

listSep(graph, X, Y, Z = integer(0), R, check = TRUE)

allSep(graph, X, Z = integer(0), R, check = TRUE)

whSep(graph, X, Z, check = TRUE)
}
\arguments{
\item{graph}{a summary graph of class \code{mixedgraph}}

\item{X, Y}{sets to separate}

\item{Z}{set to include in conditioning set}

\item{R}{maximal set to look in}

\item{check}{logical: check that the sets are suitably disjoint?}
}
\description{
Implement the algorithms of van der Zander et al. (2014) to find separator sets.
}
\details{
\code{findSep} finds a set that m-separates \code{X} and \code{Y}, 
that must contain vertices in \code{Z} and may contain vertices in \code{R}.  
\code{listSep} exhaustively lists all such sets.  

\code{allSep} returns the maximal set m-separated from \code{X} by 
\code{Z}. \code{whSep} returns the subset of \code{Z} that is m-separated 
from \code{X} by the remainder of \code{Z}.

The algorithms for \code{findSep} and \code{listSep} are taken
from van der Zander et al. (2014).
}
\section{Functions}{
\itemize{
\item \code{listSep()}: list all m-separations

\item \code{allSep()}: give maximal m-separated set

\item \code{whSep()}: find m-separated subset

}}
\references{
van der Zander et al. Constructing Separators and Adjustment Sets in Ancestral Graphs.
\emph{UAI Workshop on Causal Inference: Learning and Prediction}, 2014.
}
