% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitADMG.R
\name{fitADMG}
\alias{fitADMG}
\alias{print.mixed_fit}
\title{Fit data to an ADMG model}
\usage{
fitADMG(
  dat,
  graph,
  r = TRUE,
  tol = sqrt(.Machine$double.eps),
  SEs = TRUE,
  sparse = FALSE,
  quietly = TRUE,
  use_optim = TRUE
)
}
\arguments{
\item{dat}{The data, as an array of counts or a data frame whose final
column contains the counts.}

\item{graph}{An ADMG (or summary graph), as an object of class \code{graph}.}

\item{r}{Logical indicating whether or not recursive factorizations should
be used.}

\item{tol}{Numeric; if log-likelihood increases by less that \code{tol} in
one step, procedure stops.}

\item{SEs}{logical: should standard errors be calculated?}

\item{sparse}{Should sparse matrices be used?  Requires package
\code{Matrix}.}

\item{quietly}{Logical indicating whether output should be suppressed.}

\item{use_optim}{should \code{optim} be used for fitting?}
}
\value{
An object of class \code{mixed_fit}.  This is a list containing
(amongst other things): \item{params}{An object of class \code{Mparams},
containing the Moebius parameters which maximise the likelihood.}
\item{ll}{Value of the log-likelihood at the maximum.}
}
\description{
Fit discrete data to the Markov structure implied by an acyclic directed
mixed graph or summary graph.
}
\details{
Fits data using coordinate-wise block descent algorithm, with an Armijo line
search.  Details in Evans and Richardson (2010).

Recursive factorizations allow for modelling of Verma-constraints.  See
Richardson et al (2012).
}
\section{Warning}{
 For the algorithm to be guaranteed to work correctly, all
counts for marginal tables consisting of districts and their parents should
be positive.  A warning will be produced if this is not so.
}

\examples{

data(gss_small)
data(gr1, package="MixedGraphs")

out = fitADMG(gss_small, gr1)
summary(out)
# not a good fit!

}
\references{
Evans, R.J. and Richardson, T.S. (2010) - Fitting acyclic
directed mixed graphs to binary data. \emph{UAI-10}.

Richardson, T.S., Robins, J.M., Shpitser, I., Evans, R.J. (2012) -
}
\seealso{
\code{\link{summary.mixed_fit}} \code{\link{autoFit}}
}
\author{
Robin Evans, Ilya Shpitser
}
\keyword{graphs}
