% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitADMG.R
\name{fitUG}
\alias{fitUG}
\title{Fit data to an undirected graphical model}
\usage{
fitUG(dat, graph, tol = sqrt(.Machine$double.eps), SEs = TRUE, quietly = TRUE)
}
\arguments{
\item{dat}{The data, as an array of counts or a data frame whose final
column contains the counts.}

\item{graph}{An ADMG, as an object of class \code{graph}.}

\item{tol}{Numeric; if log-likelihood increases by less that \code{tol} in
one step, procedure stops.}

\item{SEs}{logical: should standard errors be calculated?}

\item{quietly}{Logical indicating whether output should be suppressed.}
}
\value{
An object of class \code{u_fit}.  This is a list containing
(amongst other things): 
\item{ll}{Value of the log-likelihood at the maximum.}
}
\description{
Fit discrete data to the Markov structure implied by an undirected graph.
}
\section{Warning}{
 For the algorithm to be guaranteed to work correctly, all
counts for marginal tables consisting of cliques and their parents should
be positive.  A warning will be produced if this is not so.
}

\references{
Lauritzen (1996), Graphical Models, OUP.
}
\author{
Robin Evans
}
\keyword{graphs}
