% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{fix}
\alias{fix}
\title{Fix a node in a CADMG}
\usage{
fix(graph, w, v)
}
\arguments{
\item{graph}{(C)ADMG object of class mixedgraph}

\item{w}{set of nodes to fix}

\item{v}{set of random nodes to retain (defaults to all not fixed)

Given a CADMG or ADMG returns the CADMG obtained by 
fixing the vertices in \code{w}.}
}
\description{
Fix a node in a CADMG
}
