% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbs.R
\name{genDist}
\alias{genDist}
\title{Generate a distribution from an ADMG model}
\usage{
genDist(graph, dims, maps, r = TRUE)
}
\arguments{
\item{graph}{an object of class graph}

\item{dims}{integer vector of dimensions of distribution}

\item{maps}{optionally, the output of \code{.getMaps()}}

\item{r}{logical indicating whether or not recursive factorizations should
be used}
}
\value{
Object of class \code{moebius} giving generalized Moebius parameters
for a distribution in the model associated with \code{graph}.
}
\description{
Generate a distribution from an ADMG model
}
