% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbs.R
\name{getProbs}
\alias{getProbs}
\title{Calculate joint probabilities.}
\usage{
getProbs(moebius, maps, graph)
}
\arguments{
\item{moebius}{Object of class \code{mparam}.}

\item{maps}{list containing maps for (use \code{.getMaps}).}

\item{graph}{if maps not supplied, an object of class \code{graph}; use to
calculate maps.}
}
\value{
A numeric array of dimensions \code{moebius$dims} giving the joint
distribution described by the Moebius parameters.
}
\description{
Given the generalized Moebius parametrization of an ADMG, calculates an
array of joint probabilities.
}
\references{
Evans and Richardson.
}
\author{
Robin Evans
}
