% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr16n18}
\alias{gr16n18}
\title{The '16-not-18' Graph}
\format{
A mixed graph object that is also an ADMG.
}
\usage{
gr16n18
}
\description{
The '16-not-18 graph' is an example of a graph with a nested constraint
the removes two parameters in the all binary case: hence the name 16 
(parameters) not 18 (parameters).
}
\keyword{datasets}
