% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr3head}
\alias{gr3head}
\title{The triskelion graph}
\format{
A maximal ancestral graph

This is a graph with a head of size three (4,5,6) such that no subset of 
size 2 is also a head.
}
\usage{
gr3head
}
\description{
The triskelion graph
}
\keyword{datasets}
