% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grCounter}
\alias{grCounter}
\title{Graph that often serves as a counter example}
\format{
A mixed graph object that is also an ADMG.
}
\usage{
grCounter
}
\description{
Graph that often serves as a counter example
}
\keyword{datasets}
