% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{group_indeps}
\alias{group_indeps}
\title{Group independences using an ordering}
\usage{
group_indeps(I)
}
\arguments{
\item{I}{a list of \code{ci} objects}
}
\description{
Group independences using an ordering
}
