% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{headsTails}
\alias{headsTails}
\title{Find heads and tails}
\usage{
headsTails(graph, r = TRUE, byDist = FALSE, intrinsic)
}
\arguments{
\item{graph}{object of class \code{graph}, should be a CADMG}

\item{r}{logical indicating if recursive heads should be used}

\item{byDist}{logical indicating if results should be returned separated by district}

\item{instrinsic}{optionally, a list of the relevant intrinsic sets of the ADMG}
}
\description{
Lists the (recursive) heads and tails for a CADMG
}
\details{
The result is a list containing elements of the 
same length named \code{heads}, \code{tails} and 
\code{intrinsic} giving the heads, tails and intrinsic sets.
If \code{byDist} is \code{TRUE} then result is a list 
each of which is of the above format, corresponding to the 
separate districts of \code{graph}.
}
