% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsicClosure}
\alias{intrinsicClosure}
\title{Find intrinsic closure of a connected set}
\usage{
intrinsicClosure(graph, set, r = TRUE, sort = 1)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{set}{set of vertices to find intrisic closure of}

\item{r}{logical: should recursive heads be used?}

\item{sort}{if sort > 1 then output is sorted}
}
\description{
Find intrinsic closure of a connected set
}
