% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsicSets}
\alias{intrinsicSets}
\title{Finds the set of all intrinsic sets of a CADMG}
\usage{
intrinsicSets(graph, r = TRUE, byDist = FALSE)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{r}{should recursive (i.e. nested) method be used?}

\item{byDist}{should results be grouped by district?}
}
\description{
Finds the set of all intrinsic sets of a CADMG
}
