% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arid.R
\name{is_ancestral}
\alias{is_ancestral}
\title{Check if a directed mixed graph is ancestral}
\usage{
is_ancestral(graph)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}
}
\description{
Check if a directed mixed graph is ancestral
}
\details{
\code{graph} should only contain directed and
bidirected edges.
}
