% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arid.R
\name{is_MAG}
\alias{is_MAG}
\title{Check if a graph is maximal and ancestral}
\usage{
is_MAG(graph)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}
}
\description{
Check if a graph is maximal and ancestral
}
\details{
Runs \code{is_ancestral} and then 
\code{is_maximal}, reports the success of both 
or the failure of either.
}
