% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arid.R
\name{is_arid}
\alias{is_arid}
\alias{is_maximal}
\alias{is_MArG}
\title{Check if a graph is maximal and arid}
\usage{
is_arid(graph)

is_maximal(graph, check = TRUE, ancestral)

is_MArG(graph, directed = FALSE)
}
\arguments{
\item{graph}{summary graph or ADMG of class \code{mixedgraph}}

\item{directed}{logical: if \code{TRUE}, undirected edges are not allowed}
}
\description{
Check if a graph is maximal and arid
}
\details{
Checks if the graph is both maximal and arid
using the functions \code{is_arid} and \code{is_maximal}.
}
\section{Functions}{
\itemize{
\item \code{is_arid}: check if a graph is arid

\item \code{is_maximal}: check if graph is maximal
}}

