% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{latentProject}
\alias{latentProject}
\title{Project out latent variables}
\usage{
latentProject(graph, latent, v, only_directed = FALSE, sort = 1)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{latent}{vertices to project out}

\item{v}{alternatively, vertices to keep}

\item{only_directed}{logical: should only directed edges be added in?}

\item{sort}{integer:1 for unique but unsorted neighbours in the \code{adjList}, 2 for sorted before being added in.}
}
\description{
Project out latent variables
}
\details{
\code{only_directed} should only be used in special situations,
such as when computing non-recursive intrisic sets.
}
