% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_connected.R
\name{m_connected}
\alias{m_connected}
\title{Get m-connected vertices}
\usage{
m_connected(graph, X, Z)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{X}{set of vertices to check for m-connection}

\item{Z}{vertices to condition on}
}
\description{
Get m-connected vertices
}
\details{
Perform a breadth-first search to find m-connected vertices given 
a specified set.

Based on the C++ function 'dConnected' from the package \code{dagitty}.
}
