% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{moebius}
\alias{moebius}
\alias{print.mparam}
\title{Evaluate generalized Moebius parameters}
\usage{
moebius(graph, ptable, dims = rep(2, n), r = TRUE)
}
\arguments{
\item{graph}{An object of class \code{mixedgraph}, an ADMG.}

\item{ptable}{An array containing a joint probability distribution over the
vertices of \code{graph}.}

\item{r}{logical - should recursive factorizations be used?}
}
\value{
An object of class \code{mparams}.  That is, a list containing:
\item{q}{The Moebius parameters.  \code{q[[d]][[i]][[j]]} is a numeric array
containing values of \eqn{P(X_H = x_H \,|\, X_T = j)} for each
\eqn{x_H}{x_H}, where \eqn{H}{H} is the \code{i}th head in district
\code{d}, and \code{j} indexes the tail states \eqn{x_T}{x_T}.}
\item{heads}{List of lists, sorted by district, and each sub-list containing
integer vectors of the heads in that district.} \item{tails}{List of lists,
containing integer vectors of the tails corresponding to the heads above.}
\item{vnames}{Vector of names of vertices in graph.} \item{dim}{Number of
categories for each variable.} \item{r}{as input}
}
\description{
Given a full probability distribution over vertices of a graph, returns the
value of the associated generalized Moebius parameters.
}
\section{Warning }{
 This function does not verify that the given distribution
satisfies the conditions of the model for the maps being calculated, and
thus the distribution will not necessarily be recovered by mapping back
using \code{probdist}.

Note this function will not generally return correct values for parameters
in the recursive parametrization unless all variables are independent.
}

\examples{

data(gr2)

# Distribution of complete independence
ptable = array(1/32, rep(2,5))
moebius(gr2, ptable, r=TRUE)

}
\references{
Evans and Richardson (2010)
}
\seealso{
\code{\link{probdist}}.
}
\author{
Robin Evans
}
