% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{partition}
\alias{partition}
\title{Return the partition function for a particular set of vertices}
\usage{
partition(graph, ht, v = graph$v, r = TRUE)
}
\arguments{
\item{graph}{a CADMG}

\item{ht}{the output of applying \code{headsTails()} to \code{graph}}

\item{v}{integer vector of vertices to partition}

\item{r}{logical indicating whether nested parameterization is being used (only if \code{ht} not provided)}
}
\description{
Return the partition function for a particular set of vertices
}
