% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{print.CADMG}
\alias{print.CADMG}
\title{Print method for a CADMG object}
\usage{
\method{print}{CADMG}(x, ...)
}
\arguments{
\item{x}{a \code{mixedgraph} that is also a \code{CADMG}}

\item{...}{other arguments to print}
}
\description{
Print method for a CADMG object
}
\details{
Prints a CADMG nicely onto the standard output
}
