% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prob_table}
\alias{prob_table}
\title{Constructs probability array for graph}
\usage{
prob_table(graph, dims = rep(2, n), values = 1/prod(dims))
}
\arguments{
\item{graph}{An ADMG, object of class \code{mixedgraph}.}

\item{dims}{An integer vector with same length as \code{graph$v} containing the number
of categories take by each random variable.}

\item{values}{A vector of joint probabilities of length \code{prod(dims)}.}
}
\value{
An array containing the probabilities with
appropriately labelled dimensions.
}
\description{
Put joint probabilities in an array with labelled dimensions.
}
\examples{

data(gr1)
prob_table(gr1)

}
\author{
Robin Evans
}
\keyword{array}
