% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdist.R
\name{probdist}
\alias{probdist}
\title{Calculate joint probabilities.}
\usage{
probdist(moebius, map, graph)
}
\arguments{
\item{moebius}{Object of class \code{mparam}.}

\item{map}{list containing maps for (use \code{maps}).}

\item{graph}{if maps not supplied, an object of class \code{graph}; use to
calculate maps.}
}
\value{
A numeric array of dimensions \code{moebius$dims} giving the joint
distribution described by the Moebius parameters.
}
\description{
Given the generalized Moebius parametrization of an ADMG, calculates an
array of joint probabilities.
}
\references{
Evans and Richardson (2010).
}
\author{
Robin Evans
}
