% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{random}
\alias{random}
\alias{fixed}
\title{Get vector of fixed or random nodes in CADMG}
\usage{
random(graph)

fixed(graph)
}
\arguments{
\item{graph}{a \code{mixedgraph} that is also a \code{CADMG}}
}
\description{
Get vector of fixed or random nodes in CADMG
}
\section{Functions}{
\itemize{
\item \code{fixed}: get fixed nodes
}}

