% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdist.R
\name{rmixedgraph}
\alias{rmixedgraph}
\title{Generate distributions from an ADMG model}
\usage{
rmixedgraph(n, graph, dims, r = TRUE, alpha = 1)
}
\arguments{
\item{n}{number of distributions to generate}

\item{graph}{an object of class graph}

\item{dims}{integer vector of dimensions of distribution}

\item{r}{logical indicating whether or not recursive factorizations should
be used}
}
\value{
Object of class \code{tables} giving distributions Markov with 
respect to a distribution in the model associated with \code{graph}.
}
\description{
Generate distributions from an ADMG model
}
\details{
The random distribution is
obtained by starting with the uniform distribution, and adding an 
independent (scaled) normal random variables to each Moebius parameter.  We 
then scale the move to reach approximately the boundary of the simplex,
and then pick a value between the uniform distribution and the boundary 
point using a Beta distribution.
}
