% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opd.R
\name{romp}
\alias{romp}
\title{Obtain the refined ordered Markov property}
\usage{
romp(graph, topOrd, power_DAG, group = FALSE)
}
\arguments{
\item{graph}{an ADMG of class \code{mixedgraph}}

\item{topOrd}{optional topological ordering}

\item{power_DAG}{a power DAG for \code{graph}}
}
\description{
Obtain the refined ordered Markov property
}
\details{
Computes a power DAG and extracts the \code{indep} attribute.
}
