% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twins}
\alias{twins}
\alias{gss}
\alias{gss_small}
\title{Data Sets}
\format{These are all \code{data.frame} objects.}
\usage{
twins

gss

gss_small
}
\description{
Some datasets for illustration.
}
\section{Functions}{
\itemize{
\item \code{gss}: another graph

\item \code{gss_small}: The instrumental variables graph
}}

\keyword{datasets}
