% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{canonicalDAG}
\alias{canonicalDAG}
\title{Canonical DAG for an ADMG}
\usage{
canonicalDAG(graph, collect = TRUE)
}
\arguments{
\item{graph}{ADMG of class \code{mixedgraph}}

\item{collect}{logical: should cliques of bidirected edges become a single latent?}
}
\description{
Canonical DAG for an ADMG
}
