% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsicClosure}
\alias{intrinsicClosure}
\title{Find intrinsic closure of a connected set}
\usage{
intrinsicClosure(graph, set, sort = 1)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{set}{set of vertices to find intrisic closure ot}

\item{sort}{}
}
\description{
Find intrinsic closure of a connected set
}
