% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arid.R
\name{is_MAG}
\alias{is_MAG}
\title{Check if a graph is maximal and ancestral}
\usage{
is_MAG(graph, directed = FALSE, .top_ord = FALSE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{directed}{logical: should this be a directed MAG?}

\item{.top_ord}{logical: if computed by ancestral should topological order be returned?}
}
\description{
Check if a graph is maximal and ancestral
}
\details{
Checks edge types, and then runs \code{is_ancestral} and 
\code{is_maximal}, reports the success of both or the failure of either.
}
