% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{localMarkovProperty}
\alias{localMarkovProperty}
\title{Local Markov property for ADMGs}
\usage{
localMarkovProperty(graph)
}
\arguments{
\item{graph}{an ADMG as an object of class \code{mixedgraph}}
}
\description{
Get a list of independences implied by the local Markov property applied to
an ADMG.
}
