% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{localMarkovProperty}
\alias{localMarkovProperty}
\title{Local Markov property for ADMGs}
\usage{
localMarkovProperty(graph, unique = TRUE, split = FALSE)
}
\arguments{
\item{graph}{an ADMG as an object of class \code{mixedgraph}}

\item{unique}{logical: should overlapping independences be removed?}

\item{split}{logical: should independences involve only singletons?}
}
\description{
Get a list of independences implied by the local Markov property applied to
an ADMG.
}
