% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGr.R
\name{rADMG}
\alias{rADMG}
\title{Function to generate random graphs}
\usage{
rADMG(n, N = 1, p_dir = 1/n, p_bi = 1/n, list = FALSE, arid = FALSE)
}
\arguments{
\item{n}{number of vertices}

\item{N}{number of graph to simulate}

\item{p_dir, p_bi}{probability of a directed/bidirected edge}

\item{list}{logical: should output always be a list?}

\item{arid}{logical: should graph be made to be arid?}
}
\description{
Function to generate random graphs
}
\details{
Uses a binomial number of edges based on \code{p_dir} and \code{p_bi}.
}
