% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdist.R, R/randomGr.R
\name{rADMG}
\alias{rADMG}
\title{Generate a distribution from an ADMG model}
\usage{
rADMG(n, N = 1, p_dir = 1/n, p_bi = 1/n, list = FALSE, arid = FALSE)

rADMG(n, N = 1, p_dir = 1/n, p_bi = 1/n, list = FALSE, arid = FALSE)
}
\arguments{
\item{n}{number of vertices}

\item{N}{number of graph to simulate}

\item{p_dir, p_bi}{probability of a directed/bidirected edge}

\item{list}{logical: should output always be a list?}

\item{arid}{logical: should graph be made to be arid?}

\item{graph}{an object of class graph}

\item{dims}{integer vector of dimensions of distribution}

\item{maps}{optionally, the output of \code{maps()}}

\item{r}{logical indicating whether or not recursive factorizations should
be used}
}
\value{
Object of class \code{moebius} giving generalized Moebius parameters
for a distribution in the model associated with \code{graph}.
}
\description{
Generate a distribution from an ADMG model
}
\details{
Uses a binomial number of edges based on \code{p_dir} and \code{p_bi}.
}
