% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdist.R
\name{rADMGdist}
\alias{rADMGdist}
\title{Generate a distribution from an ADMG model}
\usage{
rADMGdist(graph, dims, map, r = TRUE, alpha = 1)
}
\arguments{
\item{graph}{an object of class graph}

\item{dims}{integer vector of dimensions of distribution}

\item{map}{optionally, the output of \code{maps()}}

\item{r}{logical indicating whether or not recursive factorizations should
be used}

\item{alpha}{parameter to use in beta distribution}
}
\value{
Object of class \code{moebius} giving generalized Moebius parameters
for a distribution in the model associated with \code{graph}.
}
\description{
Generate a distribution from an ADMG model
}
