% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{subsetRep}
\alias{subsetRep}
\title{Get the subset representation of an ordinary ADMG}
\usage{
subsetRep(graph, max_size, sort = 1)
}
\arguments{
\item{graph}{ADMG}

\item{max_size}{largest set to consider}
}
\description{
Get the subset representation of an ordinary ADMG
}
